Add isBoolean and isString validators
