+++ src/validators/validators.ts
@@ -1,15 +1,11 @@
-export function validator(
-  validationFunction: ValidatorFunction,
-  message: String
-): Validator {
-  return (keyPath, value) => {
-    if (!validationFunction(value)) {
-      throw Error(
-        `Invalid configuration value ${value} for ${keyPath}: ${message}`
-      );
-    }
-  };
-}
+import { validator } from "./validator";
 
-type ValidatorFunction = (value: any) => boolean;
-type Validator = (keyPath: String, value: any) => void;
+export const isString = validator(
+  val => typeof val === "string",
+  "must be a string"
+);
+
+export const isBoolean = validator(
+  val => typeof val === "boolean",
+  "must be a boolean"
+);
